MAK-84/ General Application Modifications

See merge request makassed/makassed!74

+++ Backend/Makassed.Api/Services/Policies/PolicyService.cs
@@ -1,11 +1,11 @@
 ﻿using ErrorOr;
-using Makassed.Api.Repositories;
-using Makassed.Api.Models.Domain;
+using Makassed.Api.Repositories.Interfaces;
 using Makassed.Api.ServiceErrors;
 using Microsoft.IdentityModel.Tokens;
 using Sieve.Models;
 using Makassed.Api.Services.Storage;
 using Makassed.Api.Services.Users;
+using Policy = Makassed.Api.Models.Domain.Policy;
 
 namespace Makassed.Api.Services.Policies;
 
@@ -37,21 +37,37 @@ public class PolicyService : IPolicyService
         return policy is null;
     }
 
-    public async Task<List<Policy>> GetPoliciesAsync(SieveModel sieveModel, Guid chapterId)
+    private async Task<bool> CheckChapterExists(Guid id) 
+    { 
+        var chapter = await _chapterRepository.GetChapterByIdAsync(id);
+
+        return chapter is not null;
+    }
+
+    // check if policy belongs to the chapter
+    private async Task<bool> CheckPolicyBelongsToChapter(Guid chapterId, Guid policyId)
     {
+        var policy = await _policyRepository.GetPolicyByIdAsync(policyId);
+
+        return policy?.ChapterId == chapterId;
+    }
+
+    public async Task<ErrorOr<List<Policy>>> GetPoliciesAsync(SieveModel sieveModel, Guid chapterId)
+    {
+        if (!await CheckChapterExists(chapterId))
+            return Errors.Chapter.NotFound;
+
         return await _policyRepository.GetPoliciesAsync(sieveModel, chapterId);
     }
 
-    public async Task<ErrorOr<Policy>> GetPolicyByIdAsync(Guid id)
+    public async Task<ErrorOr<Policy>> GetPolicyByIdAsync(Guid chapterId, Guid id)
     {
+        if (!await CheckChapterExists(chapterId))
+            return Errors.Chapter.NotFound;
+
         var policy = await _policyRepository.GetPolicyByIdAsync(id);
         
         return policy is null ? Errors.Policy.NotFound : policy;
-    }
-
-    private async Task<Chapter?> CheckChapterExists(Guid id) 
-    { 
-        return await _chapterRepository.GetChapterByIdAsync(id);
     }
 
     public async Task<ErrorOr<Created>> CreatePolicyAsync(Guid chapterId, Policy policy)
@@ -61,9 +77,9 @@ public class PolicyService : IPolicyService
         if (userRole == null)
             return Errors.User.Unauthorized;
 
-        var existedChapterResult = await CheckChapterExists(chapterId);
+        var existedChapter = await _chapterRepository.GetChapterByIdAsync(chapterId);
 
-        if (existedChapterResult is null)
+        if (existedChapter is null)
             return Errors.Chapter.NotFound;
 
         if (!await IsUniqueName(policy.Name))
@@ -82,27 +98,37 @@ public class PolicyService : IPolicyService
         
         await _policyRepository.CreatePolicyAsync(policy);
 
-        await _chapterRepository.UpdateChapterEnableStateAsync(existedChapterResult.Id);
+        existedChapter.EnableState = (existedChapter.Policies.Count + 1) > 0;
 
         await _unitOfWork.SaveChangesAsync();
 
         return Result.Created;
     }
 
-    public async Task<ErrorOr<Deleted>> DeletePolicyAsync(Guid id)
+    public async Task<ErrorOr<Deleted>> DeletePolicyAsync(Guid chapterId, Guid id)
     {
+        var chapter = await _chapterRepository.GetChapterByIdAsync(chapterId);
+
+        if (chapter is null)
+            return Errors.Chapter.NotFound;
+
         var deletedPolicy = await _policyRepository.DeletePolicyAsync(id);
 
         if (deletedPolicy is null)
             return Errors.Policy.NotFound;
 
-        await _chapterRepository.UpdateChapterEnableStateAsync(deletedPolicy.ChapterId);
+        chapter.EnableState = (chapter.Policies.Count - 1)  > 0;
+
+        await _unitOfWork.SaveChangesAsync();
         
         return Result.Deleted;
     }
 
-    public async Task<ErrorOr<Updated>> UpdatePolicyAsync(Guid id, Policy policy)
+    public async Task<ErrorOr<Updated>> UpdatePolicyAsync(Guid chapterId, Guid id, Policy policy)
     {
+        if (await CheckPolicyBelongsToChapter(chapterId, id))
+            return Errors.Policy.DoesNotBelongToChapter;
+
         policy.PdfUrl = await _localFileStorageService.UploadFileAndGetUrlAsync(policy.MainFile);
         policy.PageCount = _localFileStorageService.GetPdfFilePageCount(policy.MainFile);
