+++ Backend/Makassed.Api/Services/PolicyDependencies/PolicyDependencyService.cs
@@ -1,13 +1,13 @@
 using ErrorOr;
 using Makassed.Api.Models.Domain;
 using Makassed.Api.Repositories;
+using Makassed.Api.Repositories.Interfaces;
 using Makassed.Api.ServiceErrors;
 using Makassed.Api.Services.Policies;
 using Makassed.Api.Services.Storage;
 using Makassed.Api.Services.Users;
 using Makassed.Contracts.Enums;
 using Sieve.Models;
-using Policy = Makassed.Api.Models.Domain.Policy;
 
 namespace Makassed.Api.Services.PolicyDependencies;
 
@@ -19,8 +19,8 @@ public class PolicyDependencyService : IPolicyDependencyService
     private readonly IUserService _userService;
 
     public PolicyDependencyService(
-        IPolicyDependencyRepository policyDependencyRepository, 
-        IPolicyService policyService, 
+        IPolicyDependencyRepository policyDependencyRepository,
+        IPolicyService policyService,
         ILocalFileStorageService localFileStorageService,
         IUserService userService)
     {
@@ -30,37 +30,48 @@ public class PolicyDependencyService : IPolicyDependencyService
         _userService = userService;
     }
 
-    private async Task<ErrorOr<Policy>> CheckExistedPolicy(Guid id)
+    // check if dependency belongs to the policy
+    private async Task<bool> CheckDependencyBelongsToPolicy(Guid policyId, Guid id)
     {
-        var findPolicyResult = await _policyService.GetPolicyByIdAsync(id);
+        var dependency = await _policyDependencyRepository.GetPolicyDependencyByIdAsync(id);
 
-        return findPolicyResult;
+        return dependency?.PolicyId == policyId;
     }
 
-    public Task<List<Dependency>> GetPolicyDependenciesAsync(Guid policyId, SieveModel sieveModel)
+    public async Task<ErrorOr<List<Dependency>>> GetPolicyDependenciesAsync(Guid chapterId, Guid policyId, SieveModel sieveModel)
     {
-        return _policyDependencyRepository.GetPolicyDependenciesAsync(policyId, sieveModel);
+        var existedPolicyResult = await _policyService.GetPolicyByIdAsync(chapterId, policyId);
+
+        if (existedPolicyResult.IsError)
+            return existedPolicyResult.Errors;
+
+        return await _policyDependencyRepository.GetPolicyDependenciesAsync(policyId, sieveModel);
     }
 
-    public Task<Dependency?> GetPolicyDependencyByIdAsync(Guid id)
+    public async Task<ErrorOr<Dependency>> GetPolicyDependencyByIdAsync(Guid policyId, Guid id)
     {
-        return _policyDependencyRepository.GetPolicyDependencyByIdAsync(id);
+        if (!await CheckDependencyBelongsToPolicy(policyId, id))
+            return Errors.PolicyDependency.DoesNotBelongToPolicy;
+
+        var policyDependency = await _policyDependencyRepository.GetPolicyDependencyByIdAsync(id);
+
+        return policyDependency is null ? Errors.PolicyDependency.NotFound : policyDependency;
     }
 
-    public async Task<ErrorOr<Dependency>> CreatePolicyDependencyAsync(Dependency policyDependency, Guid policyId)
+    public async Task<ErrorOr<Dependency>> CreatePolicyDependencyAsync(Dependency policyDependency, Guid policyId, Guid chapterId)
     {
         var userRole = await _userService.GetUserRoleAsync();
 
         if (userRole == null)
             return Errors.User.Unauthorized;
 
-        var existedPolicyResult = await CheckExistedPolicy(policyId);
+        var existedPolicyResult = await _policyService.GetPolicyByIdAsync(chapterId, policyId);
 
         if (existedPolicyResult.IsError)
             return existedPolicyResult.Errors;
 
         if (!existedPolicyResult.Value.IsApproved)
-            return Errors.PolicyDependency.CantAdd;
+            return Errors.PolicyDependency.CannotAdd;
 
         policyDependency.PolicyId = policyId;
 
@@ -78,8 +89,11 @@ public class PolicyDependencyService : IPolicyDependencyService
         return policyDependency;
     }
 
-    public async Task<ErrorOr<Deleted>> DeletePolicyDependencyAsync(Guid id)
+    public async Task<ErrorOr<Deleted>> DeletePolicyDependencyAsync(Guid policyId, Guid id)
     {
+        if (!await CheckDependencyBelongsToPolicy(policyId, id))
+            return Errors.PolicyDependency.DoesNotBelongToPolicy;
+
         var deletionResult = await _policyDependencyRepository.DeletePolicyDependencyAsync(id);
 
         return deletionResult is null ? Errors.PolicyDependency.NotFound : Result.Deleted;
@@ -92,8 +106,11 @@ public class PolicyDependencyService : IPolicyDependencyService
         return deletedPolicies is null ? Errors.PolicyDependency.NotFoundPolicyDependenciesType : deletedPolicies;
     }
 
-    public async Task<ErrorOr<Updated>> UpdatePolicyDependencyAsync(Guid id, Dependency policyDependency)
+    public async Task<ErrorOr<Updated>> UpdatePolicyDependencyAsync(Guid policyId, Guid id, Dependency policyDependency)
     {
+        if (!await CheckDependencyBelongsToPolicy(policyId, id))
+            return Errors.PolicyDependency.DoesNotBelongToPolicy;
+
         policyDependency.PdfUrl = await _localFileStorageService.UploadFileAndGetUrlAsync(policyDependency.File);
         policyDependency.PagesCount = _localFileStorageService.GetPdfFilePageCount(policyDependency.File);
